+++ extensions/cornerstone-dicom-seg/src/init.ts
@@ -1,5 +1,5 @@
 import { addTool, BrushTool } from '@cornerstonejs/tools';
 
-export default function init({ configuration = {} }): void {
+export default function init({ servicesManager }): void {
   addTool(BrushTool);
 }