+++ extensions/cornerstone/src/hps/mpr.ts
@@ -2,8 +2,10 @@ import { Types } from '@ohif/core';
 
 export const mpr: Types.HangingProtocol.Protocol = {
   id: 'mpr',
-  name: 'Multi-Planar Reconstruction',
+  name: 'MPR',
   locked: true,
+  icon: 'layout-advanced-mpr',
+  isPreset: true,
   createdDate: '2021-02-23',
   modifiedDate: '2023-08-15',
   availableTo: {},
@@ -22,11 +24,6 @@ export const mpr: Types.HangingProtocol.Protocol = {
       },
     ],
     // Turns off crosshairs when switching out of MPR mode
-    onProtocolExit: [
-      {
-        commandName: 'cleanUpCrosshairs',
-      },
-    ],
   },
   displaySetSelectors: {
     activeDisplaySet: {
@@ -90,6 +87,9 @@ export const mpr: Types.HangingProtocol.Protocol = {
                 id: 'mpr',
                 source: true,
                 target: true,
+                options: {
+                  syncColormap: true,
+                },
               },
             ],
           },
@@ -114,6 +114,9 @@ export const mpr: Types.HangingProtocol.Protocol = {
                 id: 'mpr',
                 source: true,
                 target: true,
+                options: {
+                  syncColormap: true,
+                },
               },
             ],
           },
@@ -138,6 +141,9 @@ export const mpr: Types.HangingProtocol.Protocol = {
                 id: 'mpr',
                 source: true,
                 target: true,
+                options: {
+                  syncColormap: true,
+                },
               },
             ],
           },