+++ extensions/tmtv/src/commandsModule.js
@@ -112,6 +112,7 @@ const commandsModule = ({ servicesManager, commandsManager, extensionManager }) 
       // Create a segmentation of the same resolution as the source data
       // using volumeLoader.createAndCacheDerivedVolume.
       const { viewportMatchDetails } = hangingProtocolService.getMatchDetails();
+
       const ptDisplaySet = actions.getMatchingPTDisplaySet({
         viewportMatchDetails,
       });
@@ -121,8 +122,11 @@ const commandsModule = ({ servicesManager, commandsManager, extensionManager }) 
         return;
       }
 
+      const currentSegmentations = segmentationService.getSegmentations();
+
       const segmentationId = await segmentationService.createSegmentationForDisplaySet(
-        ptDisplaySet.displaySetInstanceUID
+        ptDisplaySet.displaySetInstanceUID,
+        { label: `Segmentation ${currentSegmentations.length + 1}` }
       );
 
       // Add Segmentation to all toolGroupIds in the viewer
@@ -141,6 +145,12 @@ const commandsModule = ({ servicesManager, commandsManager, extensionManager }) 
         segmentationService.setActiveSegmentationForToolGroup(segmentationId, toolGroupId);
       }
 
+      segmentationService.addSegment(segmentationId, {
+        segmentIndex: 1,
+        properties: {
+          label: 'Segment 1',
+        },
+      });
       return segmentationId;
     },
     setSegmentationActiveForToolGroups: ({ segmentationId }) => {
@@ -150,7 +160,7 @@ const commandsModule = ({ servicesManager, commandsManager, extensionManager }) 
         segmentationService.setActiveSegmentationForToolGroup(segmentationId, toolGroupId);
       });
     },
-    thresholdSegmentationByRectangleROITool: ({ segmentationId, config }) => {
+    thresholdSegmentationByRectangleROITool: ({ segmentationId, config, segmentIndex }) => {
       const segmentation = csTools.segmentation.state.getSegmentation(segmentationId);
 
       const { representationData } = segmentation;
@@ -201,12 +211,11 @@ const commandsModule = ({ servicesManager, commandsManager, extensionManager }) 
           { volume: referencedVolume, lower: ptLower, upper: ptUpper },
           { volume: ctReferencedVolume, lower: ctLower, upper: ctUpper },
         ],
-        { overwrite: true }
+        { overwrite: true, segmentIndex }
       );
     },
-    calculateSuvPeak: ({ labelmap }) => {
+    calculateSuvPeak: ({ labelmap, segmentIndex }) => {
       const { referencedVolumeId } = labelmap;
-
       const referencedVolume = cs.cache.getVolume(referencedVolumeId);
 
       const annotationUIDs = csTools.annotation.selection.getAnnotationsSelectedByToolName(
@@ -217,7 +226,7 @@ const commandsModule = ({ servicesManager, commandsManager, extensionManager }) 
         csTools.annotation.state.getAnnotation(annotationUID)
       );
 
-      const suvPeak = calculateSuvPeak(labelmap, referencedVolume, annotations);
+      const suvPeak = calculateSuvPeak(labelmap, referencedVolume, annotations, segmentIndex);
       return {
         suvPeak: suvPeak.mean,
         suvMax: suvPeak.max,
@@ -508,13 +517,6 @@ const commandsModule = ({ servicesManager, commandsManager, extensionManager }) 
           displaySetInstanceUID: ptDisplaySet.displaySetInstanceUID,
           colormap: {
             name: colormap,
-            // TODO: This opacity mapping matches that in hpViewports, but
-            // ideally making this editable in a side panel would be useful
-            opacity: [
-              { value: 0, opacity: 0 },
-              { value: 0.1, opacity: 0.9 },
-              { value: 1, opacity: 0.95 },
-            ],
           },
         });
 
@@ -530,83 +532,51 @@ const commandsModule = ({ servicesManager, commandsManager, extensionManager }) 
   const definitions = {
     setEndSliceForROIThresholdTool: {
       commandFn: actions.setEndSliceForROIThresholdTool,
-      storeContexts: [],
-      options: {},
     },
     setStartSliceForROIThresholdTool: {
       commandFn: actions.setStartSliceForROIThresholdTool,
-      storeContexts: [],
-      options: {},
     },
     getMatchingPTDisplaySet: {
       commandFn: actions.getMatchingPTDisplaySet,
-      storeContexts: [],
-      options: {},
     },
     getPTMetadata: {
       commandFn: actions.getPTMetadata,
-      storeContexts: [],
-      options: {},
     },
     createNewLabelmapFromPT: {
       commandFn: actions.createNewLabelmapFromPT,
-      storeContexts: [],
-      options: {},
     },
     setSegmentationActiveForToolGroups: {
       commandFn: actions.setSegmentationActiveForToolGroups,
-      storeContexts: [],
-      options: {},
     },
     thresholdSegmentationByRectangleROITool: {
       commandFn: actions.thresholdSegmentationByRectangleROITool,
-      storeContexts: [],
-      options: {},
     },
     getTotalLesionGlycolysis: {
       commandFn: actions.getTotalLesionGlycolysis,
-      storeContexts: [],
-      options: {},
     },
     calculateSuvPeak: {
       commandFn: actions.calculateSuvPeak,
-      storeContexts: [],
-      options: {},
     },
     getLesionStats: {
       commandFn: actions.getLesionStats,
-      storeContexts: [],
-      options: {},
     },
     calculateTMTV: {
       commandFn: actions.calculateTMTV,
-      storeContexts: [],
-      options: {},
     },
     exportTMTVReportCSV: {
       commandFn: actions.exportTMTVReportCSV,
-      storeContexts: [],
-      options: {},
     },
     createTMTVRTReport: {
       commandFn: actions.createTMTVRTReport,
-      storeContexts: [],
-      options: {},
     },
     getSegmentationCSVReport: {
       commandFn: actions.getSegmentationCSVReport,
-      storeContexts: [],
-      options: {},
     },
     exportRTReportForAnnotations: {
       commandFn: actions.exportRTReportForAnnotations,
-      storeContexts: [],
-      options: {},
     },
     setFusionPTColormap: {
       commandFn: actions.setFusionPTColormap,
-      storeContexts: [],
-      options: {},
     },
   };