+++ platform/ui/src/components/ViewportPane/ViewportPane.tsx
@@ -64,7 +64,7 @@ function ViewportPane({
       onScroll={onInteractionHandler}
       onWheel={onInteractionHandler}
       className={classnames(
-        'hover:border-primary-light group h-full w-full overflow-hidden rounded-md transition duration-300',
+        'group h-full w-full overflow-hidden rounded-md transition duration-300',
         {
           'border-primary-light border-2': isActive,
           'border-2 border-transparent': !isActive,
@@ -77,10 +77,10 @@ function ViewportPane({
     >
       <div
         className={classnames(
-          'h-full w-full overflow-hidden rounded-md group-hover:border-transparent',
+          'h-full w-full overflow-hidden rounded-md',
           {
             'border border-transparent': isActive,
-            'border-secondary-light border': !isActive,
+            'border-secondary-light group-hover:border-primary-light/70 border': !isActive,
           },
           className
         )}