+++ platform/ui/src/components/index.js
@@ -29,7 +29,7 @@ import NavBar from './NavBar';
 import Notification from './Notification';
 import Select from './Select';
 import SegmentationTable from './SegmentationTable';
-import SegmentationGroupTable from './SegmentationGroupTable';
+import { SegmentationGroupTable, SegmentationGroupTableExpanded } from './SegmentationGroupTable';
 import SidePanel from './SidePanel';
 import SplitButton from './SplitButton';
 import StudyBrowser from './StudyBrowser';
@@ -73,16 +73,29 @@ import CheckBox from './CheckBox';
 import LoadingIndicatorProgress from './LoadingIndicatorProgress';
 import LoadingIndicatorTotalPercent from './LoadingIndicatorTotalPercent';
 import ViewportActionBar from './ViewportActionBar';
+import ViewportActionCorners, { ViewportActionCornersLocations } from './ViewportActionCorners';
 import ProgressLoadingBar from './ProgressLoadingBar';
 import LegacySidePanel from './LegacySidePanel';
 import PanelSection from './PanelSection';
 import AdvancedToolbox from './AdvancedToolbox';
 import InputDoubleRange from './InputDoubleRange';
 import LegacyButtonGroup from './LegacyButtonGroup';
+import LabellingFlow from './Labelling';
+import SwitchButton, { SwitchLabelLocation } from './SwitchButton';
+import * as AllInOneMenu from './AllInOneMenu';
+import ViewportActionArrows from './ViewportActionArrows';
+import HeaderPatientInfo from './HeaderPatientInfo';
+import LegacySplitButton from './LegacySplitButton';
+import { ToolSettings } from './AdvancedToolbox';
+import { Toolbox } from './Toolbox';
+import InvestigationalUseDialog from './InvestigationalUseDialog';
+import MeasurementItem from './MeasurementTable/MeasurementItem';
+import LayoutPreset from './LayoutPreset';
 
 export {
   AboutModal,
   AdvancedToolbox,
+  AllInOneMenu,
   HotkeyField,
   Header,
   UserPreferences,
@@ -131,6 +144,7 @@ export {
   Select,
   SegmentationTable,
   SegmentationGroupTable,
+  SegmentationGroupTableExpanded,
   SidePanel,
   SplitButton,
   StudyBrowser,
@@ -142,6 +156,8 @@ export {
   StudyListTableRow,
   StudySummary,
   Svg,
+  SwitchButton,
+  SwitchLabelLocation,
   Table,
   TableBody,
   TableCell,
@@ -157,10 +173,21 @@ export {
   TooltipClipboard,
   Typography,
   Viewport,
+  ViewportActionArrows,
   ViewportActionBar,
+  ViewportActionCorners,
+  ViewportActionCornersLocations,
   ViewportDownloadForm,
   ViewportGrid,
   ViewportPane,
   ViewportOverlay,
   WindowLevelMenuItem,
+  LabellingFlow,
+  HeaderPatientInfo,
+  LayoutPreset,
+  LegacySplitButton,
+  ToolSettings,
+  Toolbox,
+  InvestigationalUseDialog,
+  MeasurementItem,
 };